ISSUE-157: Start using readmodel where needed
