+++ src/Domain/Strava/Activity/ActivityIntensity.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Domain\Strava\Activity;
 
-use App\Domain\Strava\Activity\WriteModel\Activity;
+use App\Domain\Strava\Activity\ReadModel\ActivityDetails;
 use App\Domain\Strava\Athlete\AthleteRepository;
 use App\Domain\Strava\Ftp\FtpRepository;
 use App\Infrastructure\Exception\EntityNotFound;
@@ -17,7 +17,7 @@ final readonly class ActivityIntensity
     ) {
     }
 
-    public function calculate(Activity $activity): ?int
+    public function calculate(ActivityDetails $activity): ?int
     {
         $athlete = $this->athleteRepository->find();
         try {
