+++ src/Domain/Strava/Activity/ReadModel/DbalActivityDetailsRepository.php
@@ -1,0 +1,92 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\Strava\Activity\ReadModel;
+
+use App\Domain\Strava\Activity\ActivityId;
+use App\Domain\Strava\Activity\SportType\SportType;
+use App\Domain\Strava\Gear\GearId;
+use App\Infrastructure\Exception\EntityNotFound;
+use App\Infrastructure\Geocoding\Nominatim\Location;
+use App\Infrastructure\Serialization\Json;
+use App\Infrastructure\ValueObject\Time\SerializableDateTime;
+use Doctrine\DBAL\Connection;
+
+final class DbalActivityDetailsRepository implements ActivityDetailsRepository
+{
+    /** @var array<int|string, Activities> */
+    public static array $cachedActivities = [];
+
+    public function __construct(
+        private readonly Connection $connection,
+    ) {
+    }
+
+    public function find(ActivityId $activityId): ActivityDetails
+    {
+        $queryBuilder = $this->connection->createQueryBuilder();
+        $queryBuilder->select('*')
+            ->from('Activity')
+            ->andWhere('activityId = :activityId')
+            ->setParameter('activityId', $activityId);
+
+        if (!$result = $queryBuilder->executeQuery()->fetchAssociative()) {
+            throw new EntityNotFound(sprintf('Activity "%s" not found', $activityId));
+        }
+
+        return $this->hydrate($result);
+    }
+
+    public function findAll(?int $limit = null): Activities
+    {
+        $cacheKey = $limit ?? 'all';
+        if (array_key_exists($cacheKey, DbalActivityDetailsRepository::$cachedActivities)) {
+            return DbalActivityDetailsRepository::$cachedActivities[$cacheKey];
+        }
+
+        $queryBuilder = $this->connection->createQueryBuilder();
+        $queryBuilder->select('*')
+            ->from('Activity')
+            ->orderBy('startDateTime', 'DESC')
+            ->setMaxResults($limit);
+
+        $activities = array_map(
+            fn (array $result) => $this->hydrate($result),
+            $queryBuilder->executeQuery()->fetchAllAssociative()
+        );
+        DbalActivityDetailsRepository::$cachedActivities[$cacheKey] = Activities::fromArray($activities);
+
+        return DbalActivityDetailsRepository::$cachedActivities[$cacheKey];
+    }
+
+    public function findMostActiveState(): ?string
+    {
+        $queryBuilder = $this->connection->createQueryBuilder();
+        $queryBuilder->select("JSON_EXTRACT(location, '$.state') as state")
+            ->from('Activity')
+            ->andWhere('state IS NOT NULL')
+            ->groupBy("JSON_EXTRACT(location, '$.state')")
+            ->orderBy('COUNT(*)', 'DESC');
+
+        return $queryBuilder->executeQuery()->fetchOne();
+    }
+
+    /**
+     * @param array<string, mixed> $result
+     */
+    private function hydrate(array $result): ActivityDetails
+    {
+        $location = Json::decode($result['location'] ?? '[]');
+
+        return new ActivityDetails(
+            activityId: ActivityId::fromString($result['activityId']),
+            startDateTime: SerializableDateTime::fromString($result['startDateTime']),
+            sportType: SportType::from($result['sportType']),
+            data: Json::decode($result['data']),
+            location: $location ? Location::fromState($location) : null,
+            weather: Json::decode($result['weather'] ?? '[]'),
+            gearId: GearId::fromOptionalString($result['gearId']),
+        );
+    }
+}
