+++ src/Domain/Strava/Activity/WriteModel/ActivityRepository.php
@@ -8,19 +8,15 @@ use App\Domain\Strava\Gear\GearIds;
 
 interface ActivityRepository
 {
+    public function find(ActivityId $activityId): Activity;
+
     public function add(Activity $activity): void;
 
     public function update(Activity $activity): void;
 
     public function delete(Activity $activity): void;
 
-    public function find(ActivityId $activityId): Activity;
-
-    public function findAll(?int $limit = null): Activities;
-
     public function findActivityIds(): ActivityIds;
 
     public function findUniqueGearIds(): GearIds;
-
-    public function findMostActiveState(): ?string;
 }
